+++ src/Trait.mjs
@@ -1,7 +1,6 @@
 import { isObjectLiteral } from './isObjectLiteral.mjs';
 import { assert } from './assert.mjs';
-import { implies } from './implies.mjs';
-import { extend, variant, isData } from './index.mjs';
+import { extend, variant, variantName, isData } from './index.mjs';
 
 export const isTrait = Symbol('isTrait'),
     data = Symbol('data'),
@@ -25,25 +24,47 @@ const getAncestorFunctions = (() => {
 const protoTrait = () => { }
 protoTrait[isTrait] = true;
 protoTrait[apply] = function self(instance, ...args) {
-    if (typeof instance === 'object' && instance !== null && variant in instance) {
-        const vt = instance[variant],
-            fns = getAncestorFunctions(this),
+    let fn
+    const dataType = this[data]
+
+    if (dataType == undefined) {
+        fn = this['_']
+    } else if (dataType[isData]) {
+        if (typeof instance === 'object' && instance !== null && variant in instance)
+            fn = this[instance[variantName]]
+        else
+            throw new TypeError(`instance must be a data variant: ${String(instance)}`)
+    } else if (dataType.prototype[variant]) { // anonymous data declaration
+        if (typeof instance === 'object' && instance !== null && variant in instance) {
+            const vt = instance[variant],
+                fns = getAncestorFunctions(this);
             // have to lookup by associated variant instead of by name
-            // because a trait can be defined for an anonymous data declaration
+            // pattern matching may solve this in the future
             fn = fns.find(fn => fn[variant] === vt)
-
-        if (fn) return fn.call(this, instance, ...args);
-
-        // fallback to wildcard
-        if ('_' in this) return this['_'].call(this, instance, ...args);
-
-        throw new TypeError(`no trait defined for ${String(vt)}`)
-    } else if (this[data] == undefined) {
-        return this['_'].call(this, instance, ...args);
+        } else {
+            throw new TypeError(`instance must be a data variant: ${String(instance)}`)
+        }
+    } else if ([Number, String, Boolean, BigInt].includes(dataType)) {
+        const type = dataType,
+            typeString = type.name.toLowerCase();
+        if (typeof instance === typeString || instance instanceof type)
+            fn = this[`${instance.toString()}${typeof instance === 'bigint' ? 'n' : ''}`];
+        else
+            throw new TypeError(`instance must be a ${typeString}: ${instance}`);
     } else {
-        throw new TypeError(`instance must be a variant: ${String(instance)}`)
+        throw new TypeError(`data must be a data declaration or primitive type: ${String(dataType)}`)
     }
+
+    if (fn) return fn.call(this, instance, ...args);
+
+    // fallback to wildcard
+    if ('_' in this) return this['_'].call(this, instance, ...args);
+
+    throw new TypeError(`no trait defined for ${String(instance)}`)
 }
+
+const primitiveList = [Number, String, Boolean, BigInt],
+    msgWildcardInvalid = `Invalid Trait declaration. Wildcard '_' must be a function`;
 
 /**
  * Defines a trait for a data declaration.
@@ -59,11 +80,8 @@ export function Trait(dataDecl, traitDef) {
 
     assert(isObjectLiteral(traitDef), 'traitDef must be an object literal');
 
-    assert(implies(dataDecl == undefined, '_' in traitDef || apply in traitDef),
-        "Wildcard '_' or Symbol(apply) must be defined if dataDecl is undefined");
-
     if ("_" in traitDef) {
-        assert(typeof traitDef['_'] === 'function', `Wildcard '_' must be a function`);
+        assert(typeof traitDef['_'] === 'function', msgWildcardInvalid);
         localTraits['_'] = traitDef['_'];
     }
 
@@ -78,21 +96,23 @@ export function Trait(dataDecl, traitDef) {
 
     localTraits[data] = dataDecl
 
-    if (dataDecl == undefined) return localTraits
-
-    if (isData in dataDecl) {
+    if (dataDecl == undefined) {
+        assert('_' in traitDef || apply in traitDef,
+            "Wildcard '_' or Symbol(apply) must be defined if dataDecl is undefined");
+        return localTraits
+    } else if (isData in dataDecl) {
         if (!('_' in traitDef)) {
             // every key in dataDecl must be in traitDef
             Object.keys(dataDecl).forEach(name => {
                 assert(traitDef[name] != null, `Invalid Trait declaration. Missing definition for '${String(name)}'`);
             })
         } else {
-            assert(typeof traitDef['_'] === 'function', `Invalid Trait declaration. Wildcard '_' must be a function`);
+            assert(typeof traitDef['_'] === 'function', msgWildcardInvalid);
             localTraits['_'] = traitDef['_'];
         }
 
         // but we iterate over traitDef instead of dataDecl so we can associate the variant
-        // since the it could have override entries
+        // since it could have override entries
         for (const [name, f] of Object.entries(traitDef)) {
             assert(typeof f === 'function', `Invalid Trait declaration. '${name}' must be a function`);
             localTraits[name] = f;
@@ -105,6 +125,25 @@ export function Trait(dataDecl, traitDef) {
         assert(typeof f === 'function', `trait must be a function: ${name}`);
         localTraits[name] = f;
         f[variant] = dataDecl;
+    } else if (primitiveList.includes(dataDecl)) {
+        if (!('_' in traitDef)) {
+            if (dataDecl === Boolean) {
+                const message = `Invalid Trait declaration. Missing definition for`;
+                for (const [name, f] of Object.entries(traitDef)) {
+                    assert(typeof f === 'function', `${message} '${name}'`);
+                    localTraits[name] = f;
+                    f[variant] = undefined;
+                }
+            } else {
+                throw new TypeError(msgWildcardInvalid);
+            }
+        } else {
+            for (const [name, f] of Object.entries(traitDef)) {
+                assert(typeof f === 'function', `Invalid Trait declaration. '${name}' must be a function`);
+                localTraits[name] = f;
+                f[variant] = undefined;
+            }
+        }
     } else {
         throw new TypeError(`dataDecl must be a data declaration or a variant declaration`)
     }