+++ src/tests/point.test.mjs
@@ -1,4 +1,4 @@
-import { Data, isData, variant, Trait } from "../index.mjs"
+import { Data, isData, variant, variantName, Trait } from "../index.mjs"
 
 describe('Point tests', () => {
     const Point = Data({ Point2: ['x', 'y'], Point3: ['x', 'y', 'z'] }),
@@ -10,6 +10,8 @@ describe('Point tests', () => {
         const p2 = Point.Point2({ x: 1, y: 2 });
         expect(p2).toBeDefined();
         expect(p2[variant]).toBe(Point2);
+        expect(p2[variantName]).toBe('Point2');
+
         expect(p2.x).toBe(1);
         expect(p2.y).toBe(2);
         expect(p2.z).toBeUndefined();
@@ -17,6 +19,7 @@ describe('Point tests', () => {
         const p3 = Point.Point3({ x: 3, y: 4, z: 5 });
         expect(p3).toBeDefined();
         expect(p3[variant]).toBe(Point3);
+        expect(p3[variantName]).toBe('Point3');
         expect(p3.x).toBe(3);
         expect(p3.y).toBe(4);
         expect(p3.z).toBe(5);